lib: rename sys module
This is a simple module rename, `sysfs` -> `sys`.
